requirements.txt
asgiref==3.3.4
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
Django==3.2
django-cors-headers==3.7.0
django-filter==2.4.0
django-oauth-toolkit==1.5.0
django-oauth2-provider==0.2.6.1
djangorestframework==3.12.4
idna==2.10
jwcrypto==0.8
Markdown==3.3.4
oauthlib==3.1.0
psycopg2-binary==2.8.6
pycparser==2.20
pytz==2021.1
requests==2.25.1
shortuuid==1.0.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.4
